LayoutTests/imported/w3c/web-platform-tests/content-security-policy/script-src-attr-elem/script-src-attr-allowed-src-blocked-expected.txt

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'script-src-attr'.

CONSOLE MESSAGE: Refused to execute a script for an inline event handler because 'unsafe-inline' does not appear in the script-src directive of the Content Security Policy.


Harness Error (TIMEOUT), message = null

NOTRUN Should not fire a security policy violation event
